refactor: split BundleFacts value/persistence/orchestration (ADR-061 gap E)
Closes the last unstarted slice of ADR-061 gap E: bundle_facts.py,
bundle_facts_serialization.py, and bundle_facts_store.py were all
classified `workflows`, conflating the BundleFacts value type, its
persistence, and capture/comparison orchestration in one layer.

Applies the same split Phase 5 already proved for *_metadata.py:

- model/bundle_facts.py: the BundleFacts value type, its schema-version
  constants, and its construction invariant
  (require_degraded_members_known).
- storage/bundle_facts_codec.py: JSON (de)serialization
  (bundle_facts_to_dict/from_dict, load/save_bundle_facts,
  looks_like_bundle_facts_document).
- storage/bundle_facts_archive.py: the G40 content-addressed zip
  archive format.
- storage/bundle_facts_package.py: the multi-artifact ProjectSnapshot
  package adapter (ADR-062 A1.4/A1.5), reclassified `storage` now that
  its only dependencies (model.bundle_facts, storage.bundle_facts_codec,
  storage.import_bundle_facts, ...) are all storage-legal.
- workflows/bundle_facts_capture.py: capture_bundle_facts,
  bundle_snapshot_from_facts.
- workflows/bundle_facts_compare.py: compare_bundle_from_facts.

The three original flat modules are now delegation-only compatibility
facades (registered in architecture/modules.yaml's `facades` list,
each under the 150-line cap with an explicit __all__), so every
existing `from abicheck.bundle_facts import ...` call site — including
the documented Python API (docs/use/multi-binary.md) and the test
suite — is unaffected. Internal callers now import the canonical owner
directly (ADR-061 D6): bundle_multibuild.py, bundle_side_input.py,
cli_compare_release_helpers.py, storage/import_bundle_facts.py,
storage/variant_composition.py, workflows/bundle_compare_operand.py,
workflows/bundle_stored_pair_compare.py, workflows/release_scope.py.

The one real serialization.py <-> storage.bundle_facts_codec two-file
cycle (each needs a name from the other) is kept dynamic via
importlib.import_module, same shape as before the move and not a new
bridge — no IMPORT_CYCLE_ALLOWLIST entry added.

Legacy-reader behavior (schema-version fallback, the degraded_members
marker, the G40 archive format, resource-limit hardening) is pinned by
the existing tests/test_bundle_facts*.py suite, which passes unchanged
before and after the move; two files' monkeypatch-based tests were
updated to patch the real owner module instead of the facade (D10).

Verified: mypy abicheck/, ruff check, scripts/check_architecture.py,
scripts/check_ai_readiness.py all clean; the full bundle_facts/
bundle_side_input/cli_compare_release/release_scope test slice (603
tests) passes.
