+++ abicheck/serialization.py
@@ -24,7 +24,7 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any, TypeVar, cast
 
 if TYPE_CHECKING:
-    from .bundle_facts import BundleFacts
+    from .model.bundle_facts import BundleFacts
     from .snapshot_io import SnapshotWriteResult
 from . import qualified_name_segments
 from .errors import IncompatibleSnapshotSchemaError, SnapshotError
@@ -1482,16 +1482,17 @@ def write_snapshot(
     )
 
 
-# ADR-061: BundleFacts (de)serialization moved to bundle_facts_serialization.py
-# (classified `workflows`, alongside the `BundleFacts` it serializes) --
-# bundle_facts.py is itself at its own 800-line production cap, so this is a
-# new sibling rather than growing that module. Each wrapper below resolves
-# its implementation via `importlib.import_module` (a runtime call, not a
-# static `ast.Import`/`ast.ImportFrom` node) rather than a
-# `from .bundle_facts_serialization import ...` -- that module itself needs
+# ADR-061 gap E: BundleFacts (de)serialization is classified `storage`
+# (`storage.bundle_facts_codec`), alongside every other snapshot/baseline
+# codec -- moved out of the flat `bundle_facts_serialization.py` facade
+# module, itself moved out of the historical flat `bundle_facts.py` before
+# that module had a settled ADR-061 layer at all. Each wrapper below still
+# resolves its implementation via `importlib.import_module` (a runtime call,
+# not a static `ast.Import`/`ast.ImportFrom` node) rather than a
+# `from .storage.bundle_facts_codec import ...` -- that module itself needs
 # `snapshot_to_dict`/`snapshot_from_dict` from *this* module, and a static
 # import in both directions is exactly the `serialization <->
-# bundle_facts_serialization` cycle `scripts/check_ai_readiness.py`'s
+# storage.bundle_facts_codec` cycle `scripts/check_ai_readiness.py`'s
 # `import-cycle-growth` check flags via a full `ast.walk` (so even a
 # function-scoped `from ... import ...` counts) -- the same reason
 # `abicheck.cli`'s own `__getattr__` resolves its moved names through
@@ -1506,7 +1507,7 @@ def write_snapshot(
 def _bundle_facts_serialization() -> Any:
     import importlib
 
-    return importlib.import_module(".bundle_facts_serialization", __package__)
+    return importlib.import_module(".storage.bundle_facts_codec", __package__)
 
 
 def bundle_facts_to_dict(facts: BundleFacts) -> dict[str, Any]:
