+++ abicheck/workflows/release_scope.py
@@ -67,8 +67,8 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from ..bundle_facts import BundleFacts
     from ..bundle_manifest import InstantiationManifest
+    from ..model.bundle_facts import BundleFacts
     from ..model.package_inventory import PackageInventory
 
 from ..model.scope_acquisition import (
